refactor: reorganize ASR and LLM adapters
