+++ packages/adapters/src/asr/adapter/index.ts
@@ -1,0 +1,1 @@
+export * from "./create-asr-provider-adapter";
