+++ packages/adapters/src/asr/index.ts
@@ -1,5 +1,5 @@
 export * from "./adapter";
 export * from "./client";
 export * from "./errors";
-export * from "./provider";
+export * from "./ports";
 export * from "./types";
