+++ packages/adapters/src/llm/adapter/index.ts
@@ -1,0 +1,1 @@
+export * from "./create-llm-provider-adapter";
