+++ packages/adapters/src/llm/client/index.ts
@@ -1,0 +1,1 @@
+export * from "./openai";
