+++ packages/adapters/src/llm/client/openai.ts
@@ -1,18 +1,22 @@
 import OpenAI from "openai";
-import type {
-  ChatCompletionAssistantMessageParam,
-  ChatCompletionMessageParam,
-  ChatCompletionSystemMessageParam,
-  ChatCompletionToolMessageParam,
-  ChatCompletionUserMessageParam,
-} from "openai/resources/chat/completions";
 import { z } from "zod";
 
 import type { Logger } from "@english-app/observability";
 
-import { type LLMClient, type LLMClientCallOptions, LLMClientError } from "../client";
 import {
-  type ChatMessage,
+  buildCompletionParams,
+  type CompletionConfig,
+  type CompletionParams,
+  createEvaluateAnswerPrompt,
+  createGeneratePlanPrompt,
+  createInterviewRubricPrompt,
+  DEFAULT_MAX_TOKENS,
+  DEFAULT_TEMPERATURE,
+  toChatCompletionMessage,
+} from "../mappers/openai";
+import type { LLMClient, LLMClientCallOptions } from "../ports/client";
+import { LLMClientError } from "../ports/client";
+import {
   type ChatReply,
   type ChatReplyInput,
   type EvaluateAnswerInput,
@@ -87,23 +91,10 @@ const interviewRubricEvalSchema = z.object({
   ),
 });
 
-const DEFAULT_TEMPERATURE = 0.7;
-const DEFAULT_MAX_TOKENS = 3000;
-
-export interface OpenAiLLMClientConfig {
+export interface OpenAiLLMClientConfig extends CompletionConfig {
   apiKey: string;
-  model: string;
   baseURL?: string;
-  temperature?: number;
-  maxTokens?: number;
   logger?: Logger;
-}
-
-interface CompletionParams {
-  model: string;
-  messages: ChatCompletionMessageParam[];
-  temperature: number;
-  maxTokens?: number;
 }
 
 function parseJsonResponse<T>(
@@ -195,157 +186,6 @@ function handleOpenAiError(error: unknown, context: string, logger?: Logger): ne
     code: "unknown_error",
     cause: error,
   });
-}
-
-function toChatCompletionMessage(message: ChatMessage): ChatCompletionMessageParam {
-  const common = message.metadata ? { metadata: message.metadata } : {};
-
-  switch (message.role) {
-    case "system":
-      return {
-        role: "system",
-        content: message.content,
-        ...(message.name ? { name: message.name } : {}),
-        ...common,
-      } satisfies ChatCompletionSystemMessageParam;
-    case "user":
-      return {
-        role: "user",
-        content: message.content,
-        ...(message.name ? { name: message.name } : {}),
-        ...common,
-      } satisfies ChatCompletionUserMessageParam;
-    case "assistant":
-      return {
-        role: "assistant",
-        content: message.content,
-        ...(message.name ? { name: message.name } : {}),
-        ...common,
-      } satisfies ChatCompletionAssistantMessageParam;
-    case "tool": {
-      const toolCallId =
-        typeof message.metadata?.tool_call_id === "string"
-          ? message.metadata.tool_call_id
-          : undefined;
-
-      if (!toolCallId) {
-        throw new LLMClientError({
-          message: "Tool messages require metadata.tool_call_id",
-          code: "invalid_request",
-        });
-      }
-
-      return {
-        role: "tool",
-        content: message.content,
-        tool_call_id: toolCallId,
-        ...common,
-      } satisfies ChatCompletionToolMessageParam;
-    }
-    default: {
-      const exhaustiveCheck: never = message.role;
-      throw new LLMClientError({
-        message: `Unsupported chat role: ${exhaustiveCheck}`,
-        code: "invalid_request",
-      });
-    }
-  }
-}
-
-function createGeneratePlanPrompt(input: GeneratePlanInput) {
-  const systemPrompt =
-    "You are an expert ESL curriculum designer creating interview-focused study plans for Brazilian software professionals. " +
-    "Produce concise, actionable lesson plans that follow the provided JSON contract.";
-
-  const userPrompt = [
-    "Learner details:",
-    JSON.stringify(
-      {
-        learner: input.learner,
-        timeframeWeeks: input.timeframeWeeks,
-        priorKnowledge: input.priorKnowledge,
-        preferences: input.preferences,
-        locale: input.locale ?? "pt-BR",
-      },
-      null,
-      2,
-    ),
-    "Requirements:",
-    "- Provide weekly objectives aligned with interview preparation.",
-    "- Ensure lesson durations sum close to the learner's available time.",
-    "- Tailor tone and recommendations considering locale and goals.",
-  ].join("\n");
-
-  return { systemPrompt, userPrompt };
-}
-
-function createEvaluateAnswerPrompt(input: EvaluateAnswerInput) {
-  const systemPrompt =
-    "You are an experienced English interviewer assessing responses from Brazilian tech professionals. " +
-    "Return structured feedback in JSON using CEFR-inspired grading.";
-
-  const userPrompt = [
-    "Question:",
-    input.question,
-    "",
-    "Candidate answer:",
-    input.answer,
-    "",
-    input.expectedAnswer ? `Reference answer:\n${input.expectedAnswer}\n` : "",
-    "Evaluation criteria:",
-    JSON.stringify(input.evaluationCriteria, null, 2),
-    "",
-    "Locale:",
-    input.locale ?? "pt-BR",
-  ]
-    .filter(Boolean)
-    .join("\n");
-
-  return { systemPrompt, userPrompt };
-}
-
-function createInterviewRubricPrompt(input: InterviewRubricEvalInput) {
-  const systemPrompt =
-    "You are an interview coach scoring candidate transcripts against a rubric. " +
-    "Provide evidence-based scoring and suggestions in JSON.";
-
-  const userPrompt = [
-    "Transcript:",
-    input.transcript,
-    "",
-    "Rubric:",
-    JSON.stringify(input.rubric, null, 2),
-    "",
-    "Context:",
-    JSON.stringify(input.context ?? {}, null, 2),
-  ].join("\n");
-
-  return { systemPrompt, userPrompt };
-}
-
-function buildCompletionParams(
-  prompts: { systemPrompt: string; userPrompt: string },
-  config: OpenAiLLMClientConfig,
-  options?: LLMClientCallOptions,
-): CompletionParams {
-  const metadataTemperature = options?.metadata?.temperature;
-  const parsedTemperature =
-    typeof metadataTemperature === "string" ? Number.parseFloat(metadataTemperature) : undefined;
-
-  const effectiveTemperature =
-    Number.isFinite(parsedTemperature) && parsedTemperature !== undefined
-      ? (parsedTemperature as number)
-      : (config.temperature ?? DEFAULT_TEMPERATURE);
-
-  return {
-    model: config.model,
-    temperature: effectiveTemperature,
-    maxTokens: config.maxTokens ?? DEFAULT_MAX_TOKENS,
-    messages: [
-      { role: "system", content: prompts.systemPrompt },
-      { role: "user", content: prompts.userPrompt },
-    ],
-  };
 }
 
 async function runJsonCompletion<T>(
