+++ packages/adapters/src/llm/index.ts
@@ -1,6 +1,6 @@
 export * from "./adapter";
 export * from "./client";
 export * from "./errors";
-export * from "./provider";
-export * from "./providers/openai-client";
+export * from "./mappers";
+export * from "./ports";
 export * from "./types";
